requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
Werkzeug==3.1.3
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9

# HTTP & API
requests==2.31.0
httpx==0.28.1
aiohttp==3.9.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# ML (lightweight)
scikit-learn==1.3.0
joblib==1.3.2

# API Connectors
openai==1.6.1
google-generativeai==0.3.2
alpha-vantage==2.3.1

# Authentication
PyJWT==2.8.0

# Text Processing
textblob==0.17.1

# Environment
python-dotenv==1.0.0

# Async
asyncio-throttle==1.0.2
websockets==12.0

# System monitoring
psutil==5.9.7

# Image processing (lightweight)
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
tenacity==8.2.3

# Telegram notifications (optional)
python-telegram-bot==20.7

# Charting
matplotlib==3.8.2
